screengrabber: fix devicePixelRatio detection
At least on two of my devices, `QApplication::desktop()->devicePixelRatio()` is always 0. This patch fixes detection of the `devicePixelRatio` by fetching the correct value from the screen that the current desktop is on, and also corrects the calculation of the selection area in `screengrabber` when the pixel ratio of the pixmap is set correctly.

The bug causes no problems on non-HiDPI devices or devices with integer scale factors for some reason that I haven't investigated yet (maybe it has a default value of 1 if 0 is fed into `setDevicePixelRatio`?). However, on my laptop with a scale factor of `1.5` (set through KDE System Settings on Xorg), flameshot breaks and the background is scaled improperly.