README: add repology badge inside collapsed div (#422)
Using https://repology.org/metapackage/flameshot/badges it's possible for users to see what version of flameshot are available in the package ecosystem. This badge is nested in an expandable div so it doesn't overwhelm the README with its size. This commit also fixes a minor typo
FYI HTML was used instead of markdown due to the markdown not being displayed.