Undo fallback mechanism for entropy gathering
This is actually unnecessary since RAND_bytes will have been seeded
from the system entropy pool automatically. If it fail, it would be
due to not having enough entropy in the system.

Because Qt's global RNG instance also seeds from the system, it will
also fail (or produce inadequate output) due to the same problem.

So instead, go the safe route and fail automatically if OpenSSL fails
to generate appropriate random bytes for the key.