Real fix for scale ratio lupoDharkael/flameshot#678
We don't need to resize pixmap as it is rendered correctly with devicePixelRatio detected by QT.

I took decision to inline setScaledPixmap as we do not need to mess around with image scale,
and this scaling thing is crippling image pixels due to approximation.

My first approach was to use `size() / scale` instead of `size()*scale` it actually worked, but was distorting pixels etc.

This seems more clear solution.