Use ${CMAKE_INSTALL_PREFIX} to provide APP_PREFIX
Currently the value of $prefix is hardcoded in the source code.
This is not a good choice given we have switched to CMake and
that CMake provides the builtin variable ${CMAKE_INSTALL_PREFIX}
to provide this information.

This patch uses the value of ${CMAKE_INSTALL_PREFIX} to populate
flameshot's internal variable APP_PREFIX as well as using it
in generating the org.flameshot.Flameshot.service file with
a proper $prefix value (via CMAKE_INSTALL_FULL_BINDIR as provided
by GNUInstallDirs).