Improve pixelation of small areas
Previously the level of pixelation would decrease if the selected area
was too thin to hold 10 pixelated blocks. Now the level of pixelation
only decreases when the selected area is too thin to hold even one
pixelated block.

This doesn't completely solve the issue of pixelation decreasing when
the area is thin, but it significantly improves the situation.