Add a configurable keyboard shortcut to commit text.
This new shortcut defaults to `ctrl+enter`, and it commits the text
you're currently typing. It does *not* change the behavior of enter
(pressing enter still inserts a newline), nor does it actually exit
flameshot. This isn't exactly what was requested, but I'm going to
somewhat cavalierly say that this fixes
https://github.com/flameshot-org/flameshot/issues/812.

*Note: I added a new user-visible string here, but I didn't update any
ts files. Is that ok?