Fix for aspect ratio losing ratio (bad code)
The problem was that the ratio was being calculated on every "tick" of movement instead of keeping it.
So every pixel moved, a new ratio was calculated which is not the user's intent.
This way, the ratio is kept while the user is resizing the image and tries to minimize the rounding issues.
Rounding issues still exist but they are now only minimal on at what is highest of the ratio.