Fix for side panel widget out of screen #3522
The problem was that the panel was moved relative to (0, 0) coordinates,
which is not an issue in case of single display, or multiple displays
in the same row.
Since capture widget is placed on all screens, the point (0, 0) will be
the top left corner of virtual screen instead of top left corner of the
current screen concluding from QDesktopWidget documentation (*).
This caused the panel to be shifted up if the height of the primary
screen was smaller than the virtual one.

 * https://doc.qt.io/qt-5/qdesktopwidget.html#screen-geometry

Signed-off-by: Petru Gurduza <petrugurduza1@gmail.com>