Disable grims png compression
Grim by default used level 6 for compressing PDFs, which could take as
long as a second depending on the hardware. This caused a big delay when
taking a screenshot with grim, leading to a bad user experience.
This commit sets  the compression level to 0(uncompressed) radically
speeding up the process of taking a screenshot.