Bypass the Qt GNOME/Wayland workaround on fixed Qt versions
We implement code on GNOME desktops to force the
QT_QPA_PLATFORM to be 'xcb'; this works around a
clipboard-related bug on GNOME+Wayland+Qt.

This bug was fixed (or worked around) in Qt 5.15.2,
so we implement a version check; if the runtime Qt
version is < 5.15.2, still force the workaround;
otherwise, we don't need the workaround so we skip
it.