Document build steps for working translations
Flameshot uses CMAKE_INSTALL_PREFIX to set APP_PREFIX.
APP_PREFIX is used to determine one of the potential locations of the translations.
This can be observed in PathInfo::translationsPaths, src/utils/pathinfo.cpp:26.
CMAKE_INSTALL_PREFIX defaults to /usr/local on UNIX.

The problem is that APP_PREFIX is set during the build -S step and changing this
using build --install build --prefix ... does not actually change the value of
APP_PREFIX. This leads to the translations not being found when running
from a custom install location.

To combat this, the --install-prefix argument is used to properly set
the APP_PREFIX. This requires CMAKE 3.21. Also possible is to use
-DCMAKE_INSTALL_PREFIX:PATH=<INSTALL_DIR> which would not require a
newer CMAKE.

Alternatives:
- Get rid of APP_PREFIX and try to find the translations using path traversal.
- Simplify the commands by using the CMAKE_INSTALL_PREFIX env var,
  which was added in 3.29 and allows getting rid of --install-prefix and --prefix.
  https://cmake.org/cmake/help/latest/variable/CMAKE_INSTALL_PREFIX.html
- Add a symbolic link next to the binary and use the same method used
  to make running from the build directory possible.
  ln -s "$INST_DIR/bin/translations" ../share/flameshot/translations