Fix English translations not used despite being preferred
When specifying the preferred language on Linux, the $LANGUAGE env var
is used. https://www.gnu.org/software/gettext/manual/html_node/The-LANGUAGE-variable.html

Example:
$ LANGUAGE=en:nl flameshot

Expected:
flameshot should be in English as this is a localization that flameshot provides.
If English would not be supported, it should fall back to using Dutch (nl).

Actual:
flameshot uses Dutch translations.

Cause:
If an empty translation file for "en" is not provided, English is only
used as a fallback.
See https://bugreports.qt.io/browse/QTBUG-69196