Add support for logging to a log file in the $XDG_STATE_DIR/flameshot (or equivelent directories on windows/mac-os)
By default logging to file is enabled for all log messages.

Filtering log messages written to disk by level severity and choosing a
different path to save the log files is configurable via the config menu.

There is also a shortcut on the about menu popup to open the configured
logging directory. This shortcut is disabled if logging to file is
disabled

The log file implements a simple roller that saves only the two most
recent sets of logs generated by flameshot. This should stop log
messages from eating users' disk.

We use cmake to check the version of QT6. If greater than or equal to
6.7, we set the default log directory to the standard App State
directory. If not, we use the app local data directory