Fix Wayland screenshot scaling on multi-monitor setups
The newly added if condition enables correct screenshot scaling in
CaptureWidget when running the app with the XCB platform plugin.
Previously the else branch would be taken, which would apply incorrect
scaling. Since the screenshot may already be scaled, we need to
convert it back to pysical dimensions and compare it with desktop
geometry again.